general: rename `src` to `pkg`
This seems to be more idiomatic in Go projects.
