+++ pkg/conference/conference.go
@@ -17,9 +17,9 @@ limitations under the License.
 package conference
 
 import (
-	"github.com/matrix-org/waterfall/src/common"
-	"github.com/matrix-org/waterfall/src/peer"
-	"github.com/matrix-org/waterfall/src/signaling"
+	"github.com/matrix-org/waterfall/pkg/common"
+	"github.com/matrix-org/waterfall/pkg/peer"
+	"github.com/matrix-org/waterfall/pkg/signaling"
 	"github.com/pion/webrtc/v3"
 	"github.com/sirupsen/logrus"
 	"maunium.net/go/mautrix/event"
