+++ pkg/conference/config.go
