+++ scripts/profile.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 
-go run ./src/*.go --cpuProfile cpuProfile.pprof --memProfile memProfile.pprof --logTime
+go run ./pkg/*.go --cpuProfile cpuProfile.pprof --memProfile memProfile.pprof --logTime
